+++ packages/cli/src/commands/network/parameters-l2.test.ts
@@ -1,4 +1,3 @@
-import { CeloDistributionScheduleWrapper } from '@celo/contractkit/lib/wrappers/CeloDistributionScheduleWrapper'
 import { testWithAnvilL2 } from '@celo/dev-utils/lib/anvil-test'
 import { stripAnsiCodesFromNestedArray, testLocallyWithWeb3Node } from '../../test-utils/cliUtils'
 import Parameters from './parameters'
@@ -7,62 +6,33 @@ process.env.NO_SYNCCHECK = 'true'
 
 testWithAnvilL2('network:parameters', (web3) => {
   test('runs', async () => {
-    jest
-      .spyOn(CeloDistributionScheduleWrapper.prototype, 'getConfig')
-      .mockImplementation(async () => {
-        return {
-          carbonOffsetting: {
-            fraction: '1000000000000000000000',
-            partner: '0x0000000000000000000000000000000000000003',
-          },
-          communityReward: {
-            fraction: '250000000000000000000000',
-            fund: '0xA2e328097D5805e37414fbF80F43D71c9B588F23',
-          },
-          distributableAmount: '409271032214310717281708',
-          targetCeloDistribution: {
-            carbonFundTargetRewards: '1630561881331915208293',
-            communityTargetRewards: '407640470332978802073415',
-            targetCeloDistribution: '612754949032214310717281708',
-          },
-        }
-      })
-
     const spy = jest.spyOn(console, 'log')
     await testLocallyWithWeb3Node(Parameters, [], web3)
     expect(stripAnsiCodesFromNestedArray(spy.mock.calls)).toMatchInlineSnapshot(`
       [
         [
-          "CeloDistributionSchedule: 
-        carbonOffsetting: 
-          fraction: 1000000000000000000000
-          partner: 0x0000000000000000000000000000000000000003
-        communityReward: 
-          fraction: 250000000000000000000000
-          fund: 0xA2e328097D5805e37414fbF80F43D71c9B588F23
-        distributableAmount: 409271032214310717281708
-        targetCeloDistribution: 
-          carbonFundTargetRewards: 1630561881331915208293
-          communityTargetRewards: 407640470332978802073415
-          targetCeloDistribution: 612754949032214310717281708
-      DowntimeSlasher: 
+          "DowntimeSlasher: 
         slashableDowntime: 5 minutes
         slashingIncentives: 
           penalty: 100000000000000000000 (~1.000e+20)
           reward: 10000000000000000000 (~1.000e+19)
       Election: 
-        currentThreshold: 0 
+        currentThreshold: 60000000000000000000 (~6.000e+19)
         electabilityThreshold: 0.001 
         electableValidators: 
           max: 100 
           min: 1 
         maxNumGroupsVotedFor: 10 
-        totalVotes: 0 
+        totalVotes: 60000000000000000000000 (~6.000e+22)
+      EpochManager: 
+        currentEpochNumber: 4
+        epochDuration: 86400
+        isTimeForNextEpoch: false
       FeeCurrencyDirectory: 
         intrinsicGasForAlternativeFeeCurrency: 
-          0x0c6a0fde0A72bA3990870f0F99ED79a821703474: 21000 (~2.100e+4)
-          0x603931FF5E63d2fd3EEF1513a55fB773d8082195: 21000 (~2.100e+4)
-          0x82398F079D742F9D0Ae71ef8C99E5c68b2eD6705: 21000 (~2.100e+4)
+          0x06f60E083aDf016a98E3c7A1aFfa1c097B617aB9: 21000 (~2.100e+4)
+          0x9cA64d4663B4A623C3E9a7F9155451647592bEc7: 21000 (~2.100e+4)
+          0xC458f5ab25a47741205722d465cDea9aB1E1154A: 21000 (~2.100e+4)
       Governance: 
         concurrentProposals: 3 
         dequeueFrequency: 4 hours
@@ -77,12 +47,12 @@ testWithAnvilL2('network:parameters', (web3) => {
           Execution: 1 week
           Referendum: 1 day
       LockedGold: 
-        totalLockedGold: 60000000000000000000000 (~6.000e+22)
+        totalLockedGold: 120000000000000000000000 (~1.200e+23)
         unlockingPeriod: 6 hours
       Reserve: 
         frozenReserveGoldDays: 0 
         frozenReserveGoldStartBalance: 0 
-        frozenReserveGoldStartDay: 19927 (~1.993e+4)
+        frozenReserveGoldStartDay: 19991 (~1.999e+4)
         otherReserveAddresses: 
 
         tobinTaxStalenessThreshold: 3153600000 (~3.154e+9)
@@ -94,7 +64,7 @@ testWithAnvilL2('network:parameters', (web3) => {
         groupLockedGoldRequirements: 
           duration: 6 months
           value: 10000000000000000000000 (~1.000e+22)
-        maxGroupSize: 5 
+        maxGroupSize: 2 
         membershipHistoryLength: 60 
         slashingMultiplierResetPeriod: 1 month
         validatorLockedGoldRequirements: