+++ packages/cli/src/commands/transfer/celo.test.ts
@@ -196,9 +196,9 @@ testWithAnvilL1('transfer:celo cmd', (web3: Web3) => {
       )
     ).rejects.toThrowErrorMatchingInlineSnapshot(`
       "0x1234567890123456789012345678901234567890 is not a valid fee currency. Available currencies:
-      0x0c6a0fde0A72bA3990870f0F99ED79a821703474 - Celo Euro (Celo Euro)
-      0x603931FF5E63d2fd3EEF1513a55fB773d8082195 - Celo Brazilian Real (Celo Brazilian Real)
-      0x82398F079D742F9D0Ae71ef8C99E5c68b2eD6705 - Celo Dollar (Celo Dollar)"
+      0x06f60E083aDf016a98E3c7A1aFfa1c097B617aB9 - Celo Euro (cEUR)
+      0x9cA64d4663B4A623C3E9a7F9155451647592bEc7 - Celo Dollar (cUSD)
+      0xC458f5ab25a47741205722d465cDea9aB1E1154A - Celo Brazilian Real (cREAL)"
     `)
   })
 })