+++ packages/dev-utils/src/anvil-test.ts
@@ -20,6 +20,12 @@ export const STABLES_ADDRESS = '0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266'
 // Introducing a different name for the same address to avoid confusion
 export const DEFAULT_OWNER_ADDRESS = STABLES_ADDRESS
 
+export const GROUP_ADDRESSES = [
+  '0x70997970C51812dc3A010C7d01b50e0d17dc79C8',
+  '0x3C44CdDdB6a900fa2b585dd299e03d12FA4293BC',
+  '0x90F79bf6EB2c4f870365E785982E1f101E93b906',
+]
+
 // Addresses are generated during migrations and need to be extracted from logs
 export enum LinkedLibraryAddress {
   AddressSortedLinkedListWithMedian = '0x5fbdb2315678afecb367f032d93f642f64180aa3',