+++ packages/sdk/contractkit/src/contract-cache.ts
@@ -8,10 +8,10 @@ import { Web3ContractCache } from './web3-contract-cache'
 import { AccountsWrapper } from './wrappers/Accounts'
 import { AttestationsWrapper } from './wrappers/Attestations'
 import { BlockchainParametersWrapper } from './wrappers/BlockchainParameters'
-import { CeloDistributionScheduleWrapper } from './wrappers/CeloDistributionScheduleWrapper'
 import { DoubleSigningSlasherWrapper } from './wrappers/DoubleSigningSlasher'
 import { DowntimeSlasherWrapper } from './wrappers/DowntimeSlasher'
 import { ElectionWrapper } from './wrappers/Election'
+import { EpochManagerWrapper } from './wrappers/EpochManager'
 import { EpochRewardsWrapper } from './wrappers/EpochRewards'
 import { Erc20Wrapper } from './wrappers/Erc20Wrapper'
 import { EscrowWrapper } from './wrappers/Escrow'
@@ -26,6 +26,7 @@ import { LockedGoldWrapper } from './wrappers/LockedGold'
 import { MultiSigWrapper } from './wrappers/MultiSig'
 import { OdisPaymentsWrapper } from './wrappers/OdisPayments'
 import { ReserveWrapper } from './wrappers/Reserve'
+import { ScoreManagerWrapper } from './wrappers/ScoreManager'
 import { SortedOraclesWrapper } from './wrappers/SortedOracles'
 import { StableTokenWrapper } from './wrappers/StableTokenWrapper'
 import { ValidatorsWrapper } from './wrappers/Validators'
@@ -33,7 +34,6 @@ import { ValidatorsWrapper } from './wrappers/Validators'
 const WrapperFactories = {
   [CeloContract.Accounts]: AccountsWrapper,
   [CeloContract.BlockchainParameters]: BlockchainParametersWrapper,
-  [CeloContract.CeloDistributionSchedule]: CeloDistributionScheduleWrapper,
   [CeloContract.EpochRewards]: EpochRewardsWrapper,
   [CeloContract.ERC20]: Erc20Wrapper,
   [CeloContract.Escrow]: EscrowWrapper,
@@ -48,6 +48,7 @@ const WrapperFactories = {
   [CeloContract.MultiSig]: MultiSigWrapper,
   [CeloContract.OdisPayments]: OdisPaymentsWrapper,
   [CeloContract.Reserve]: ReserveWrapper,
+  [CeloContract.ScoreManager]: ScoreManagerWrapper,
   [CeloContract.StableToken]: StableTokenWrapper,
   [CeloContract.StableTokenEUR]: StableTokenWrapper,
   [CeloContract.StableTokenBRL]: StableTokenWrapper,
@@ -62,6 +63,7 @@ const WrapperFactoriesWhichNeedCache = {
   [CeloContract.DoubleSigningSlasher]: DoubleSigningSlasherWrapper,
   [CeloContract.DowntimeSlasher]: DowntimeSlasherWrapper,
   [CeloContract.Election]: ElectionWrapper,
+  [CeloContract.EpochManager]: EpochManagerWrapper,
   [CeloContract.Governance]: GovernanceWrapper,
   [CeloContract.LockedGold]: LockedGoldWrapper,
   [CeloContract.Validators]: ValidatorsWrapper,
@@ -81,10 +83,10 @@ interface WrapperCacheMap {
   [CeloContract.Accounts]?: AccountsWrapper
   [CeloContract.Attestations]?: AttestationsWrapper
   [CeloContract.BlockchainParameters]?: BlockchainParametersWrapper
-  [CeloContract.CeloDistributionSchedule]?: CeloDistributionScheduleWrapper
   [CeloContract.DoubleSigningSlasher]?: DoubleSigningSlasherWrapper
   [CeloContract.DowntimeSlasher]?: DowntimeSlasherWrapper
   [CeloContract.Election]?: ElectionWrapper
+  [CeloContract.EpochManager]?: EpochManagerWrapper
   [CeloContract.EpochRewards]?: EpochRewardsWrapper
   [CeloContract.ERC20]?: Erc20Wrapper<IERC20>
   [CeloContract.Escrow]?: EscrowWrapper
@@ -101,6 +103,7 @@ interface WrapperCacheMap {
   // [CeloContract.Random]?: RandomWrapper,
   // [CeloContract.Registry]?: RegistryWrapper,
   [CeloContract.Reserve]?: ReserveWrapper
+  [CeloContract.ScoreManager]?: ScoreManagerWrapper
   [CeloContract.SortedOracles]?: SortedOraclesWrapper
   [CeloContract.StableToken]?: StableTokenWrapper
   [CeloContract.StableTokenEUR]?: StableTokenWrapper
@@ -136,9 +139,6 @@ export class WrapperCache implements ContractCacheType {
   getBlockchainParameters() {
     return this.getContract(CeloContract.BlockchainParameters)
   }
-  getCeloDistributionSchedule() {
-    return this.getContract(CeloContract.CeloDistributionSchedule)
-  }
   getDoubleSigningSlasher() {
     return this.getContract<CeloContract.DoubleSigningSlasher>(CeloContract.DoubleSigningSlasher)
   }
@@ -150,6 +150,9 @@ export class WrapperCache implements ContractCacheType {
   }
   getEpochRewards() {
     return this.getContract(CeloContract.EpochRewards)
+  }
+  getEpochManager() {
+    return this.getContract(CeloContract.EpochManager)
   }
   getErc20(address: string) {
     return this.getContract(CeloContract.ERC20, address)
@@ -190,6 +193,9 @@ export class WrapperCache implements ContractCacheType {
   }
   getReserve() {
     return this.getContract(CeloContract.Reserve)
+  }
+  getScoreManager() {
+    return this.getContract(CeloContract.ScoreManager)
   }
   getSortedOracles() {
     return this.getContract(CeloContract.SortedOracles)